Struct redlock::redlock::Redlock

pub struct Redlock {
    // some fields omitted
}

Distributed Lock Manager class object

Methods

impl Redlock

fn dlm(urls: Vec<String>, retry_count: Option<i32>, retry_delay: Option<f32>) -> RedlockResult<Redlock>

instantialize a lock manager with a vector of URLs (format: redis://host:port/db) for each Redis master node, and optionally specify a retry count and delay between each retry. If None is given, the default will be 3 and 0.2s respectively

fn lock(&mut self, res_name: String, ttl: i32) -> RedlockResult<Lock>

locks resource specified by res_name for ttl in miliseconds

fn unlock(&mut self, lock: Lock) -> RedlockResult<()>

unlocks resource held by Lock